requirements.txt
streamlit
paho-mqtt
pandas
numpy
fastapi
uvicorn[standard]
python-multipart
Pillow
librosa
soundfile
scikit-learn
opencv-python-headless